# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amiga Specialty HR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training:
    get:
      operationId: listAllTrainingRecords
      tags: [training]
      summary: List training records across all employees
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, expired, overdue, expiring_soon, all]
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Records with employee info
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrainingRecordWithEmployee" }
...
  /training/employees/{id}:
    get:
      operationId: listEmployeeTrainingRecords
      tags: [training]
      summary: List training records for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrainingRecord" }
    post:
      operationId: createTrainingRecord
      tags: [training]
      summary: Add training record to employee
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTrainingRecordBody" }
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingRecord" }
  /training-records/{recordId}:
    patch:
      operationId: updateTrainingRecord
      tags: [training]
      summary: Update training record
      parameters:
        - name: recordId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTrainingRecordBody" }
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingRecord" }
    delete:
      operationId: deleteTrainingRecord
      tags: [training]
      summary: Delete training record
      parameters:
        - name: recordId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    TrainingRecord:
      type: object
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        name: { type: string }
        category:
          type: string
          enum: [mandatory, professional, technical, soft_skills, compliance, other]
        provider: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, in_progress, completed, expired]
        cost: { type: ["number", "null"] }
        currency: { type: string }
        certificateObjectPath: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        isMandatory: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, employeeId, name, category, status, currency, isMandatory, createdAt, updatedAt]
    TrainingRecordWithEmployee:
      allOf:
        - $ref: "#/components/schemas/TrainingRecord"
        - type: object
          properties:
            employeeName: { type: string }
            employeeNumber: { type: string }
            department: { type: string }
          required: [employeeName, employeeNumber, department]
    CreateTrainingRecordBody:
      type: object
      properties:
        name: { type: string }
        category:
          type: ["string", "null"]
          enum: [mandatory, professional, technical, soft_skills, compliance, other, null]
        provider: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        status:
          type: ["string", "null"]
          enum: [pending, in_progress, completed, expired, null]
        cost: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        isMandatory: { type: ["boolean", "null"] }
      required: [name]
    UpdateTrainingRecordBody:
      type: object
      properties:
        name: { type: string }
        category:
          type: string
          enum: [mandatory, professional, technical, soft_skills, compliance, other]
        provider: { type: ["string", "null"] }
        completedDate: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, in_progress, completed, expired]
        cost: { type: ["number", "null"] }
        currency: { type: string }
        notes: { type: ["string", "null"] }
...